Moved queries in util under data
